# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProjetBank API",
        "description": "API de gestion bancaire",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/v1/users/{userId}/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Clients d'un utilisateur",
                "description": "Get clients by user ID",
                "operationId": "4bb454370418dd2b7c0c5d66ba48928a",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des clients de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "uuid"
                                            },
                                            "user_id": {
                                                "type": "string",
                                                "example": "uuid"
                                            },
                                            "numeroCompte": {
                                                "type": "string",
                                                "example": "CB12345678"
                                            },
                                            "titulaire": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "type": {
                                                "type": "string",
                                                "enum": [
                                                    "cheque",
                                                    "epargne"
                                                ]
                                            },
                                            "solde": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1000.5
                                            },
                                            "devise": {
                                                "type": "string",
                                                "example": "FCFA"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "enum": [
                                                    "actif",
                                                    "bloque"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/clients/{id}/comptes": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Comptes d'un client",
                "description": "Get client comptes",
                "operationId": "d1cd2da93e58f6d8e22f5de8b4715a63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes du client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "uuid"
                                            },
                                            "client_id": {
                                                "type": "string",
                                                "example": "uuid"
                                            },
                                            "numeroCompte": {
                                                "type": "string",
                                                "example": "CB12345678"
                                            },
                                            "solde": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1000.5
                                            },
                                            "devise": {
                                                "type": "string",
                                                "example": "FCFA"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "enum": [
                                                    "actif",
                                                    "bloque"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé"
                    }
                }
            }
        },
...
    },
...
}